third_party/WebKit/LayoutTests/css3/device-adapt/viewport-initial-width-zero-expected.txt

ALERT: viewport size 100x200 scale -1 with limits [-1, -1] and userScalable true
